refactor: Split  into smaller components
