+++ .changeset/refactor-dashboard-filters-modal.md
@@ -1,0 +1,5 @@
+---
+'@hyperdx/app': patch
+---
+
+refactor: Split `DashboardFiltersModal` into smaller components
