+++ packages/app/src/components/DashboardFiltersModal/CustomInputWrapper.tsx
@@ -1,0 +1,39 @@
+import { FieldError } from 'react-hook-form';
+import { Input, Tooltip } from '@mantine/core';
+import { IconInfoCircle } from '@tabler/icons-react';
+
+interface CustomInputWrapperProps {
+  children: React.ReactNode;
+  label: string;
+  tooltipText?: string;
+  error?: FieldError;
+}
+
+export const CustomInputWrapper = ({
+  children,
+  label,
+  tooltipText,
+  error,
+}: CustomInputWrapperProps) => {
+  const errorMessage =
+    error &&
+    (error.message ||
+      (error?.type === 'required' ? 'This field is required' : 'Error'));
+
+  return (
+    <div>
+      <Input.Label>{label}</Input.Label>
+      {tooltipText && (
+        <Tooltip label={tooltipText}>
+          <IconInfoCircle size={14} className="ms-2" />
+        </Tooltip>
+      )}
+      {errorMessage && (
+        <Input.Error color="red" size="sm">
+          {errorMessage}
+        </Input.Error>
+      )}
+      <div className="mt-1">{children}</div>
+    </div>
+  );
+};
