+++ packages/app/src/components/DashboardFiltersModal/constants.ts
@@ -1,0 +1,1 @@
+export const MODAL_SIZE = 'lg';
